Renames package load
