--- next/load/doc.go
@@ -1,2 +1,0 @@
-// Package load loads data to be persisted
-package load
