+++ next/loadtransactions/doc.go
@@ -1,0 +1,2 @@
+// Package loadtransactions loads data to be persisted
+package loadtransactions
